+++ bin/core/terminal.js
@@ -1,11 +1,11 @@
-const package = require('../../package')
+const applicationConfiguration = require('../../package')
 const _ = require('lodash')
 
 module.exports = {
   registerVorpalCommand: (vorpal, configuration) => {
     const isInteractive = process.argv.length <= 2
-    const majorVersion = parseInt(_.head(package.version.split(`.`)), 10)
-    const applicationHandle = `${package.name}-${majorVersion}`
+    const majorVersion = parseInt(_.head(applicationConfiguration.version.split(`.`)), 10)
+    const applicationHandle = `${applicationConfiguration.name}-${majorVersion}`
 
     vorpal
       .delimiter(`>> `)