+++ bin/packages/folder_package.js
@@ -1,8 +1,5 @@
 const fs = require('fs-extra')
 
-const { configuration } = require('../core/configuration')
-const paths = configuration.get(`paths`)
-
 class FolderPackage {
   static sendToStaging ({ stagingPath }, cachePath) {
     return new Promise((resolve, reject) => {