+++ bin/packages/tar_package.js
@@ -1,9 +1,6 @@
 const FileSystem = require('../support/file_system')
 const tar = require('tar')
 
-const { configuration } = require('../core/configuration')
-const paths = configuration.get(`paths`)
-
 class TarPackage {
   static sendToStaging ({ stagingPath }, cachePath) {
     return new Promise((resolve, reject) => {