+++ test/strategies/reference_strategy_spec.js
@@ -5,11 +5,10 @@ const ReferenceStrategy = require('../../bin/strategies/reference_strategy')
 const { configuration } = require(`../../bin/core/configuration`)
 
 test.cb('strategy: reference (folder)', t => {
-  t.plan(6)
+  t.plan(5)
 
   const {
     stagingPath,
-    archivePath,
     addendum,
     version,
     archive,
@@ -22,7 +21,6 @@ test.cb('strategy: reference (folder)', t => {
   t.is(uri, `../folder/`, `uri is the base path to the resource and needs the addendum to resolve the full path`)
 
   t.is(stagingPath, `${configuration.get('paths.staging')}/${archive}/`, `staging path is the archive name in staging folder`)
-  t.is(archivePath, `${configuration.get('paths.archives')}/${archive}/`, `archive path is the archive name in staging folder`)
 
   t.end()
 })