+++ changelog.md
@@ -3,7 +3,8 @@
 ## 0.2.2
 
 - REFACTORED command prompt in interactive mode in Vorpal
--
+- REFACTORED expand staging folder fully to avoid collisions
+- FIXED read component name from manifest on archive or fallback to folder name without version
 
 ## 0.2.1